LayoutTests/editing/selection/ios/select-text-in-container-with-backdrop-filter.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true SelectionHonorsOverflowScrolling=true ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width,initial-scale=1">
    <script src="../../../resources/ui-helper.js"></script>
    <script src="../../../resources/js-test.js"></script>
    <style>
    html, body {
        font-family: system-ui;
        font-size: 16px;
    }

    section {
        width: 100%;
        height: 100px;
        text-align: center;
        backdrop-filter: blur(24px);
        background-color: rgba(200, 200, 200, 0.1);
        border-bottom: 1px solid gray;
        line-height: 100px;
        position: absolute;
        top: 0;
    }

    #target {
        border: 1px solid tomato;
        border-radius: 4px;
        padding: 4px;
        font-size: 24px;
    }
    </style>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        description("This test verifies that the native text selection renders correctly in containers with a CSS backdrop filter.");

        const target = document.getElementById("target");

        await UIHelper.longPressElement(target);
        await UIHelper.waitForSelectionToAppear();

        const {x, y} = UIHelper.midPointOfRect(await UIHelper.selectionBounds());
        viewWithSelection = await UIHelper.frontmostViewAtPoint(x, y);

        getSelection().removeAllRanges();
        await UIHelper.ensurePresentationUpdate();
        viewWithoutSelection = await UIHelper.frontmostViewAtPoint(x, y);

        shouldNotBe("viewWithSelection", "viewWithoutSelection");
        finishJSTest();
    });
    </script>
</head>
<body>
    <section>Select <span id="target">this</span> text.</section>
</body>
</html>
